[Fix] database path

+++ database.py
@@ -8,7 +8,8 @@ import shutil
 DB_PATH = os.path.expanduser("~/apps/autoboot/database.sqlite")
 
 # copy paste database to local dir
-shutil.copyfile("data/database.sqlite", DB_PATH)
+if not os.path.isfile(DB_PATH):
+    shutil.copyfile("data/database.sqlite", DB_PATH)
 
 
 class RTDatabase():
@@ -17,7 +18,7 @@ class RTDatabase():
         pass
         
     def open_database(self):
-        self.DB_connection = sqlite3.connect("data/database.sqlite")
+        self.DB_connection = sqlite3.connect(DB_PATH)
         self.DB_cursor = self.DB_connection.cursor()
     
     def close_database(self):
